# chat/docs/swagger.yaml
basePath: /v1
definitions:
...
  friend.BlackParams:
    properties:
      black:
        description: 是否拉黑
        example: 1
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - user_id
    type: object
  friend.DestroyParams:
    properties:
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - user_id
    type: object
  friend.RemarkParams:
    properties:
      nickname:
        description: 备注内侧
        example: test
        type: string
      tags:
        description: 标签
        example:
        - test
        - test1
        items:
          type: string
        type: array
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - nickname
    - user_id
    type: object
  friend.StarParams:
    properties:
      star:
        description: 是否星标用户
        example: 1
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - user_id
    type: object
  group.ActionParams:
    properties:
      id:
        description: 群ID
        example: 1
        type: integer
      user_id:
        description: 用户ID
        example: 1
        type: integer
    required:
    - id
    - user_id
    type: object
  group.NicknameParams:
    properties:
      id:
        description: 群ID
        example: 1
        type: integer
      nickname:
        description: 群名
        example: name
        type: string
    required:
    - id
    - nickname
    type: object
  group.UpdateParams:
    properties:
      id:
        description: 群ID
        example: 1
        type: integer
      name:
        description: 群名
        example: name
        type: string
      remark:
        description: 群公告
        example: remark
        type: string
    required:
    - id
    type: object
...
  user.PhoneLoginParams:
    properties:
      phone:
        description: 手机号
        example: 13333333333
        type: integer
      verify_code:
        description: 验证码
        example: "888888"
        type: string
    required:
    - phone
    - verify_code
    type: object
...
host: 127.0.0.1:9050
info:
  contact:
    email: test@test.com
    name: test
    url: http://www.swagger.io/support
  description: chat api
  title: chat docs api
  version: "1.0"
paths:
...
  /apply/friend:
    post:
      description: 申请好友
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 好友ID
        in: body
        name: friend_id
        required: true
        schema:
          type: integer
      - description: 备注昵称
        in: body
        name: nickname
        required: true
        schema:
          type: string
      - description: 看我
        in: body
        name: look_me
        required: true
        schema:
          type: integer
      - description: 看他
        in: body
        name: look_him
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6Ik"}}'
          schema:
            type: string
      summary: 申请好友
      tags:
      - 好友申请
...
  /chat/detail:
    post:
      consumes:
      - application/json
      description: 获取聊天信息
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 用户/群组id
        in: body
        name: id
        required: true
        schema:
          type: integer
      - description: 类型，1=用户，2=群组
        in: body
        name: type
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 获取聊天信息
      tags:
      - 聊天
...
  /friend/auth:
    post:
      consumes:
      - application/json
      description: 删除好友
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: destroy
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/friend.DestroyParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 删除好友
      tags:
      - 好友
  /friend/black:
    post:
      consumes:
      - application/json
      description: 加入/移除黑名单
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: black
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/friend.BlackParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 加入/移除黑名单
      tags:
      - 好友
...
  /friend/remark:
    post:
      consumes:
      - application/json
      description: 设置备注标签
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      - description: remark
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/friend.RemarkParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 设置备注标签
      tags:
      - 好友
  /friend/star:
    post:
      consumes:
      - application/json
      description: 加入/移除星标
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      - description: star
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/friend.StarParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 加入/移除星标
      tags:
      - 好友
...
  /group/edit:
    post:
      consumes:
      - application/json
      description: 更新群组信息
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: The group info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/group.UpdateParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 更新群组信息
      tags:
      - 群组
...
  /group/kickoff:
    post:
      description: 踢出群成员
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: The group info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/group.ActionParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{}}'
          schema:
            type: string
      summary: 踢出群成员
      tags:
      - 群组
...
  /group/nickname:
    post:
      consumes:
      - application/json
      description: 更新群昵称
      parameters:
      - description: The group info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/group.NicknameParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 更新群昵称
      tags:
      - 群组
...
  /group/user:
    get:
      consumes:
      - application/json
      description: 获取群成员
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 群ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":null}'
          schema:
            type: string
      summary: 获取群成员
      tags:
      - 群组
  /login:
    post:
      description: 通过用户名密码登录
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6Ik"}}'
          schema:
            type: string
      summary: 用户登录接口
      tags:
      - 用户
  /login_phone:
    post:
      description: 仅限手机登录
      parameters:
      - description: phone
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.PhoneLoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6Ik"}}'
          schema:
            type: string
      summary: 用户登录接口
      tags:
      - 用户
  /logout:
    get:
      description: 用户注销登录
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6Ik"}}'
          schema:
            type: string
      summary: 用户注销登录
      tags:
      - 用户
...
  /moment/timeline:
    get:
      description: 我的朋友圈
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 页码
        in: query
        name: p
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"token":"eyJhbGciOiJIUzI1NiIsInR5cCI6Ik"}}'
          schema:
            type: string
      summary: 我的朋友圈
      tags:
      - 朋友圈
...
  /user/search:
    get:
      consumes:
      - application/json
      description: 搜索用户
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 搜索关键词
        in: body
        name: keyword
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"msg":"OK","data":{}}'
          schema:
            type: string
      summary: 搜索用户
      tags:
      - 用户
...
swagger: "2.0"